更新代码，修复虚拟实体控制器中没有分页总数和数据总数问题，为此增加了PageBaseController类

+++ Xin.Entity/VirtualEntity/UsTagTypeInventory.cs
@@ -13,6 +13,6 @@ namespace Xin.Entities.VirtualEntity
         public string TagType { get; set; }
         public int Qty { get; set; }
 
-        public int RowNumber { get; set; }
+        public long RowNumber { get; set; }
     }
 }
